feat: Split Spring Integrations into Modules - Meeds-io/MIPs#57 This change will allow to split Spring integrations with Kernel & Portal, so that addons can choose which integrations to apply.
