+++ component/api/src/main/java/org/exoplatform/commons/api/persistence/ExoEntityProcessor.java
@@ -23,6 +23,8 @@ import java.util.Set;
 import javax.annotation.processing.AbstractProcessor;
 import javax.annotation.processing.RoundEnvironment;
 import javax.annotation.processing.SupportedAnnotationTypes;
+import javax.annotation.processing.SupportedSourceVersion;
+import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.TypeElement;
 import javax.tools.Diagnostic;
@@ -42,6 +44,7 @@ import org.apache.commons.collections.CollectionUtils;
  * 7/22/15
  */
 @SupportedAnnotationTypes("org.exoplatform.commons.api.persistence.ExoEntity")
+@SupportedSourceVersion(SourceVersion.RELEASE_17)
 public class ExoEntityProcessor extends AbstractProcessor {
 
   /**
