+++ component/common/src/main/java/io/meeds/spring/web/security/FilterConfiguration.java
@@ -13,19 +13,22 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.exoplatform.commons.persistence.impl;
+package io.meeds.spring.web.security;
 
-import org.exoplatform.commons.InitialContextInitializer2;
-import org.exoplatform.container.xml.InitParams;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
-public class LiquibaseDataInitializerTestContext extends LiquibaseDataInitializer {
+@Configuration
+public class FilterConfiguration {
 
-  public LiquibaseDataInitializerTestContext(InitialContextInitializer2 initialContextInitializer, InitParams initParams) {
-    super(initialContextInitializer, initParams);
+  @Bean
+  public FilterRegistrationBean<PortalIdentityFilter> identityFilter() {
+    FilterRegistrationBean<PortalIdentityFilter> registrationBean = new FilterRegistrationBean<>();
+    registrationBean.setFilter(new PortalIdentityFilter());
+    registrationBean.addUrlPatterns("/rest/*");
+    registrationBean.setOrder(1);
+    return registrationBean;
   }
 
-  @Override
-  public void start() { // NOSONAR
-    super.start();
-  }
 }
