+++ component/common/src/main/java/io/meeds/spring/web/security/PortalIdentityFilter.java
@@ -1,6 +1,8 @@
-/*
+/**
  * This file is part of the Meeds project (https://meeds.io/).
- * Copyright (C) 2020 - 2022 Meeds Association contact@meeds.io
+ *
+ * Copyright (C) 2020 - 2023 Meeds Association contact@meeds.io
+ *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
@@ -9,23 +11,20 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * Lesser General Public License for more details.
+ *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.exoplatform.commons.persistence.impl;
 
-import org.exoplatform.commons.InitialContextInitializer2;
-import org.exoplatform.container.xml.InitParams;
+package io.meeds.spring.web.security;
 
-public class LiquibaseDataInitializerTestContext extends LiquibaseDataInitializer {
+import org.exoplatform.services.security.web.SetCurrentIdentityFilter;
 
-  public LiquibaseDataInitializerTestContext(InitialContextInitializer2 initialContextInitializer, InitParams initParams) {
-    super(initialContextInitializer, initParams);
-  }
+/**
+ * A Web filter to retrieve currently authenticated user Identity to current Web
+ * context session
+ */
+public class PortalIdentityFilter extends SetCurrentIdentityFilter {
 
-  @Override
-  public void start() { // NOSONAR
-    super.start();
-  }
 }
