+++ component/test/core/src/main/java/org/exoplatform/component/test/BaseGateInTest.java
@@ -70,4 +70,5 @@ public class BaseGateInTest extends TestCase {
         afe.initCause(t);
         return afe;
     }
+
 }
